onUnitComplete()/onUnitRenegade()/onSaveGame() are now timed correctly
Fix the function signature for onUnitComplete() (the signature changed in BWAPI v4.0.0), so that it is considered in the timing logic.
Time onUnitRenegade() and onSaveGame(). I think these callbacks were overlooked in the initial implementation of StarcraftAITournamentManager. Note that the timing of onSaveGame() doesn't include the time to actually save the file (that happens before the bot's onSaveGame() is called).
(No impact) Remove onPlayerDropped() because it was removed from BWAPI several versions ago.
(No impact) Use 'override' specifier where appropriate (this is a safe change).